information boxes refactor to popups - create mixin to pass help_text to template
