+++ epilepsy12/models/assessment.py
@@ -1,6 +1,8 @@
 from django.db import models
 
-from ..general_functions import calculate_time_elapsed
+from epilepsy12.models.help_text_mixin import HelpTextMixin
+
+from ..general_functions import *
 from ..constants import CAN_ONLY_VIEW_GENERAL_PAEDIATRIC_CENTRE, CAN_ONLY_VIEW_TERTIARY_NEUROLOGY_CENTRE, CAN_ONLY_VIEW_CHILDRENS_EPILEPSY_SURGERY_CENTRE, CAN_ALLOCATE_GENERAL_PAEDIATRIC_CENTRE, CAN_UPDATE_GENERAL_PAEDIATRIC_CENTRE, CAN_DELETE_GENERAL_PAEDIATRIC_CENTRE, CAN_ALLOCATE_TERTIARY_NEUROLOGY_CENTRE, CAN_EDIT_TERTIARY_NEUROLOGY_CENTRE, CAN_DELETE_TERTIARY_NEUROLOGY_CENTRE, CAN_CONFIRM_CHILDRENS_EPILEPSY_SURGICAL_SERVICE_REFERRAL_CRITERIA_MET, CAN_ALLOCATE_CHILDRENS_EPILEPSY_SURGERY_CENTRE, CAN_EDIT_CHILDRENS_EPILEPSY_SURGERY_CENTRE, CAN_DELETE_CHILDRENS_EPILEPSY_SURGERY_CENTRE
 from .time_and_user_abstract_base_classes import *
 
@@ -8,7 +10,7 @@ from .time_and_user_abstract_base_classes import *
 from .registration import Registration
 
 
-class Assessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
+class Assessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass, HelpTextMixin):
     """
     This class stores information on each assessment performed during the registration period.
     Each Case has only a single initial assessment (the first)
