+++ epilepsy12/models/episode.py
@@ -1,12 +1,15 @@
 from tabnanny import verbose
 from django.db import models
 from django.contrib.postgres.fields import ArrayField
+
+from epilepsy12.models.help_text_mixin import HelpTextMixin
 from .multiaxial_diagnosis import MultiaxialDiagnosis
 from ..constants import DATE_ACCURACY, EPISODE_DEFINITION, EPILEPSY_DIAGNOSIS_STATUS, EPILEPSY_SEIZURE_TYPE, NON_EPILEPSY_SEIZURE_TYPE, NON_EPILEPSY_SEIZURE_ONSET, NON_EPILEPTIC_SYNCOPES, NON_EPILEPSY_BEHAVIOURAL_ARREST_SYMPTOMS, NON_EPILEPSY_SLEEP_RELATED_SYMPTOMS, NON_EPILEPSY_PAROXYSMS, MIGRAINES, EPIS_MISC, GENERALISED_SEIZURE_TYPE
 from .time_and_user_abstract_base_classes import *
+from ..general_functions import *
 
 
-class Episode(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
+class Episode(TimeStampAbstractBaseClass, UserStampAbstractBaseClass, HelpTextMixin):
     """
     Summarises each seizure episode.
     Each child may have several seizure episodes, one of which must be epileptic to be included.
