+++ epilepsy12/models/initial_assessment.py
@@ -1,5 +1,5 @@
 from django.db import models
-from functools import partial as curry
+from epilepsy12.models.help_text_mixin import HelpTextMixin
 from ..constants import CHRONICITY, DIAGNOSTIC_STATUS
 from .time_and_user_abstract_base_classes import *
 
@@ -7,7 +7,7 @@ from .time_and_user_abstract_base_classes import *
 from .registration import Registration
 
 
-class InitialAssessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
+class InitialAssessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass, HelpTextMixin):
     """
     This class records information about the initial assessment.
     Whilst other information about the child and their epilepsy may be captured across the audit year
@@ -112,18 +112,6 @@ class InitialAssessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
         verbose_name = 'First Paediatric Assessment'
         verbose_name_plural = 'First Paediatric Assessments'
 
-    def _get_help_label_text(self, field_name):
-        """Given a field name, return it's label help text."""
-        for field in self._meta.fields:
-            if field.name == field_name:
-                return field.help_text['label']
-
-    def _get_help_reference_text(self, field_name):
-        """Given a field name, return it's reference help text."""
-        for field in self._meta.fields:
-            if field.name == field_name:
-                return field.help_text['reference']
-
     def save(
             self,
             *args, **kwargs) -> None:
@@ -134,29 +122,3 @@ class InitialAssessment(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
             return f"First Paediatric Assessment for {self.registration.case} on {self.date_of_initial_assessment}"
         else:
             return f"{self.registration.case} has not yet had First Paediatric Assessment."
-
-    def __init__(self, *args, **kwargs) -> None:
-        super(InitialAssessment, self).__init__(*args, **kwargs)
-        """
-        Thanks https://bradmontgomery.net/blog/django-hack-help-text-modal-instance/ for this snippet
-        Returns the help text methods to the template
-        Can use {{initial_assessment.get_*field*_help_label_text}} and {{initial_assessment.get_*field*_help_reference_text}}
-        in the template
-        """
-
-        # Again, iterate over all of our field objects.
-        for field in self._meta.fields:
-            # Create a string, get_FIELDNAME_help text
-            label_method_name = "get_{0}_help_label_text".format(field.name)
-            reference_method_name = "get_{0}_help_reference_text".format(
-                field.name)
-
-            # We can use curry to create the method with a pre-defined argument
-            label_curried_method = curry(
-                self._get_help_label_text, field_name=field.name)
-            reference_curried_method = curry(
-                self._get_help_reference_text, field_name=field.name)
-
-            # And we add this method to the instance of the class.
-            setattr(self, label_method_name, label_curried_method)
-            setattr(self, reference_method_name, reference_curried_method)
