+++ epilepsy12/models/registration.py
@@ -1,6 +1,7 @@
 from dateutil.relativedelta import relativedelta
 from django.db import models
 from epilepsy12.models.audit_progress import AuditProgress
+from epilepsy12.models.help_text_mixin import HelpTextMixin
 
 from .case import Case
 from ..constants import CAN_APPROVE_ELIGIBILITY, CAN_REMOVE_APPROVAL_OF_ELIGIBILITY, CAN_REGISTER_CHILD_IN_EPILEPSY12, CAN_UNREGISTER_CHILD_IN_EPILEPSY12, CAN_ONLY_VIEW_GENERAL_PAEDIATRIC_CENTRE, CAN_ALLOCATE_GENERAL_PAEDIATRIC_CENTRE, CAN_UPDATE_GENERAL_PAEDIATRIC_CENTRE, CAN_DELETE_GENERAL_PAEDIATRIC_CENTRE
@@ -8,7 +9,7 @@ from ..general_functions import *
 from .time_and_user_abstract_base_classes import *
 
 
-class Registration(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
+class Registration(TimeStampAbstractBaseClass, UserStampAbstractBaseClass, HelpTextMixin):
     """
     A record is created in the Registration class every time a case is registered for the audit
     """
