+++ epilepsy12/models/syndrome.py
@@ -1,13 +1,16 @@
 from django.db import models
 from operator import itemgetter
+
+from epilepsy12.models.help_text_mixin import HelpTextMixin
 from ..constants import SYNDROMES
 from .time_and_user_abstract_base_classes import *
+from ..general_functions import *
 
 # other tables
 from .multiaxial_diagnosis import MultiaxialDiagnosis
 
 
-class Syndrome(TimeStampAbstractBaseClass, UserStampAbstractBaseClass):
+class Syndrome(TimeStampAbstractBaseClass, UserStampAbstractBaseClass, HelpTextMixin):
     """
     This class stores information on syndromes
     One MultiaxialDescription can have multiple syndromes
